Expand Covmodel bounds to plausible values
The bound checking for the Covmodel was introduced for the variogram
estimation. The upper bounds where set to arbitrary values, which can
be confusing and cumbersum in case of having to use larger values.
The fitting algorithm of Scipy can handle np.inf as a bound, thus the
standard bounds are now set to physically plausible values, in most
cases [0, inf].