Grand Unification of struct. variogram estimators
In the cython code, merged the 1d-, 2d-, and 3d-versions of the structured
variogram estimators into the 3d version, with a tiny little bit of
preparation code in the python wrappers. Switched off the OpenMP loop
for the moment, as it always results in errors for the 1d case and
sometimes causes problems for higher dimensions. Super strange!