Add new data encapsulation and refactor
2 new classes are added in order to separate data and methods more
cleanly. This is very much WIP, but I'd like to get a discussion
started before I put in too much work.
The `FieldData` class contains information about the mesh and it
holds a dict containing the individual field values, which are
defined on that mesh, e.g. "krige" and "cond".